+++ packages/class-core-generator/src/entity-types.templates.ts
@@ -184,9 +184,9 @@ export const typeForLanguageEntity = (imports: Imports) => {
                 when(!isAbstract(classifier))(
                     () =>
                         [
-                            `static create(id: ${imports.json("LionWebId")}, handleDelta?: ${imports.generic("DeltaHandler")}, parentInfo?: ${imports.generic("Parentage")}): ${classifier.name} {`,
+                            `static create(id: ${imports.json("LionWebId")}, receiveDelta?: ${imports.generic("DeltaReceiver")}, parentInfo?: ${imports.generic("Parentage")}): ${classifier.name} {`,
                             indent([
-                                `return new ${classifier.name}(${imports.language(classifier.language)}.INSTANCE.${classifier.name}, id, handleDelta, parentInfo);`
+                                `return new ${classifier.name}(${imports.language(classifier.language)}.INSTANCE.${classifier.name}, id, receiveDelta, parentInfo);`
                             ]),
                             `}`,
                         ]
@@ -196,9 +196,9 @@ export const typeForLanguageEntity = (imports: Imports) => {
                         [
                             ``,
                             features.map(withNewlineAppended(classMembersForFeature)),
-                            `${isAbstract(classifier) ? "protected" : "public"} constructor(classifier: ${imports.core("Classifier")}, id: ${imports.json("LionWebId")}, handleDelta?: ${imports.generic("DeltaHandler")}, parentInfo?: ${imports.generic("Parentage")}) {`,
+                            `${isAbstract(classifier) ? "protected" : "public"} constructor(classifier: ${imports.core("Classifier")}, id: ${imports.json("LionWebId")}, receiveDelta?: ${imports.generic("DeltaReceiver")}, parentInfo?: ${imports.generic("Parentage")}) {`,
                             indent([
-                                `super(classifier, id, handleDelta, parentInfo);`,
+                                `super(classifier, id, receiveDelta, parentInfo);`,
                                 features.map((feature) => `this._${feature.name} = new ${imports.generic(valueManagerFor(feature))}<${tsTypeForValueManager(feature, imports)}>(${imports.language(feature.classifier.language)}.INSTANCE.${feature.classifier.name}_${feature.name}, this);`)
                             ]),
                             `}`,