+++ packages/class-core-test/src/value-managers/property_1.tests.ts
@@ -16,7 +16,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import {
-    collectingDeltaHandler,
+    collectingDeltaReceiver,
     nodeBaseDeserializer,
     PropertyAddedDelta,
     PropertyChangedDelta,
@@ -34,8 +34,8 @@ const testLanguageBase = TestLanguageBase.INSTANCE
 describe("[1] string property", () => {
 
     it("getting an unset [1] string property", () => {
-        const [handleDelta, deltas] = collectingDeltaHandler();
-        const node = DataTypeTestConcept.create("node", handleDelta);
+        const [receiveDelta, deltas] = collectingDeltaReceiver();
+        const node = DataTypeTestConcept.create("node", receiveDelta);
 
         // pre-check:
         equal(deltas.length, 0);
@@ -63,8 +63,8 @@ describe("[1] string property", () => {
         equal(nodes.length, 1);
         equal(nodes[0].properties.length, 0);
 
-        const [handleDelta, deltas] = collectingDeltaHandler();
-        const deserialize = nodeBaseDeserializer([testLanguageBase], handleDelta);
+        const [receiveDelta, deltas] = collectingDeltaReceiver();
+        const deserialize = nodeBaseDeserializer([testLanguageBase], receiveDelta);
         const problemHandler = new AccumulatingSimplisticHandler();
         const deserializedNodes = deserialize(serializationChunk, [], new BuiltinPropertyValueDeserializer(), problemHandler);
         equal(problemHandler.allProblems.length, 0);
@@ -85,8 +85,8 @@ describe("[1] string property", () => {
     });
 
     it("setting a [1] string property", () => {
-        const [handleDelta, deltas] = collectingDeltaHandler();
-        const node = DataTypeTestConcept.create("node", handleDelta);
+        const [receiveDelta, deltas] = collectingDeltaReceiver();
+        const node = DataTypeTestConcept.create("node", receiveDelta);
 
         // pre-check:
         equal(deltas.length, 0);
@@ -116,8 +116,8 @@ describe("[1] string property", () => {
             }
         );
 
-        const [handleDelta, deltas] = collectingDeltaHandler();
-        const deserialize = nodeBaseDeserializer([testLanguageBase], handleDelta);
+        const [receiveDelta, deltas] = collectingDeltaReceiver();
+        const deserialize = nodeBaseDeserializer([testLanguageBase], receiveDelta);
         const problemHandler = new AccumulatingSimplisticHandler();
         const deserializedNodes = deserialize(serializationChunk, [], new BuiltinPropertyValueDeserializer(), problemHandler);
         equal(problemHandler.allProblems.length, 0);
@@ -133,8 +133,8 @@ describe("[1] string property", () => {
     });
 
     it("unsetting a [1] string property", () => {
-        const [handleDelta, deltas] = collectingDeltaHandler();
-        const node = DataTypeTestConcept.create("node", handleDelta);
+        const [receiveDelta, deltas] = collectingDeltaReceiver();
+        const node = DataTypeTestConcept.create("node", receiveDelta);
 
         // pre-check:
         node.stringValue_1 = "foo";
@@ -152,8 +152,8 @@ describe("[1] string property", () => {
     });
 
     it("changing a [1] string property", () => {
-        const [handleDelta, deltas] = collectingDeltaHandler();
-        const node = DataTypeTestConcept.create("node", handleDelta);
+        const [receiveDelta, deltas] = collectingDeltaReceiver();
+        const node = DataTypeTestConcept.create("node", receiveDelta);
 
         // pre-check:
         node.stringValue_1 = "foo";