+++ packages/class-core/src/deltas/index.ts
@@ -18,9 +18,8 @@
 export * from "./appliers.js";
 export * from "./base.js";
 export * from "./compositor.js";
-export * from "./forwarder.js";
-export * from "./handlers.js";
 export * from "./inverters.js";
+export * from "./receivers.js";
 export * from "./types.g.js";
 export * from "./serialization/index.js";